refactor: 修复打包错误
